+++ demisto_sdk/commands/common/content/tests/objects/pack_objects/readme/readme_test.py
@@ -23,13 +23,13 @@ def test_prefix():
     assert obj.normalize_file_name() == README.name
 
 
-def test_dump():
+def test_mention_contributors_in_readme():
     """
     Given: pack README
 
     When: creating artifacts, if CONTRIBUTORS file exist in the pack
 
-    Then: Thanks to contributors
+    Then: Ensure credit to contributors is appended to the readme file
     """
     pack_path = TEST_CONTENT_REPO / PACKS_DIR / 'Sample01'
     obj = Readme(README, pack_path)